+++ now_lms/version/__init__.py
@@ -40,7 +40,7 @@ CODE_NAME = "Karla"
 # SemVer (https://semver.org)
 MAYOR = "1"
 MENOR = "1"
-PATCH = "2"
+PATCH = "3"
 
 # <--------------------------------------------------------------------------> #
 # Quick fix